[INTERNAL, TEST] Remodeling of ProteinResolver
Moved functional code from the TOPP class ProteinResolver to a helper
class ProteinResolver, which can be found in ANALYSIS/QUANTITATION.
Added a TOPP test for ProteinResolver for consensusXML files (idXML is
still missing). Created the scaffold for test class of the helper class
ProteinResolver, however, implementation is still missing. Moreover,
ProteinResolver is now included in the doxygen documentation and
registered in the ToolHandler.